backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic[email]==2.5.0
transformers>=4.40.0
cryptography==41.0.7
sentencepiece==0.1.99
python-dotenv==1.0.0
sqlalchemy==2.0.23
PyJWT>=2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
torch>=2.0.0
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0

